fix tests, fork function
